nutriAI - Copy/requirements.txt
pandas
numpy
sentence-transformers
faiss-cpu
groq
streamlit
pymongo
bcrypt
python-dotenv
streamlit-authenticator 